refactor AST and evaluator
