+++ src/core/code/ast.go
@@ -1,98 +1,12 @@
 package code
 
 import (
-	"dacrane/utils"
-	"reflect"
-	"regexp"
-	"strings"
-
 	"github.com/macrat/simplexer"
 )
 
 type Code []Entity
 
 type Entity map[string]any
-
-func (code Code) Find(kind string, name string) Entity {
-	return utils.Find(code, func(e Entity) bool {
-		return e["kind"].(string) == kind && e["name"].(string) == name
-	})
-}
-
-func (code Code) Dependency(kind string, name string) []Entity {
-	entity := code.Find(kind, name)
-	paths := references(entity)
-	var dependencies []Entity
-	for _, path := range paths {
-		identifiers := strings.Split(path, ".")
-		kind := identifiers[0]
-		name := identifiers[1]
-		dependency := code.Find(kind, name)
-		dependencies = append(dependencies, dependency)
-	}
-	return dependencies
-}
-
-func references(raw map[string]any) []string {
-	var paths []string
-	for _, v := range raw {
-		switch t := reflect.TypeOf(v); t.Kind() {
-		case reflect.Map:
-			paths = append(paths, references(v.(map[string]any))...)
-		case reflect.String:
-			r, e := regexp.Compile(`\$\{(.*?)\}`)
-			if e != nil {
-				panic(e)
-			}
-			res := r.FindAllStringSubmatch(v.(string), -1)
-			for _, exprStr := range res {
-				print(exprStr[1])
-				expr := ParseExpr(exprStr[1])
-				paths = append(paths, extractRefNames(expr)...)
-			}
-		default:
-		}
-	}
-	return paths
-}
-
-func extractRefNames(expr Expr) []string {
-	var names []string
-
-	switch e := expr.(type) {
-	case *Ref:
-		if id, ok := e.Expr.(*Identifier); ok {
-			names = append(names, id.Name)
-		}
-		names = append(names, extractRefNames(e.Expr)...)
-		names = append(names, extractRefNames(e.Key)...)
-	case *Identifier:
-		names = append(names, e.Name)
-	case *BinaryExpr:
-		names = append(names, extractRefNames(e.Left)...)
-		names = append(names, extractRefNames(e.Right)...)
-	case *UnaryExpr:
-		names = append(names, extractRefNames(e.Expr)...)
-	case *IfExpr:
-		names = append(names, extractRefNames(e.Condition)...)
-		names = append(names, extractRefNames(e.Then)...)
-		names = append(names, extractRefNames(e.Else)...)
-	case *List:
-		for _, item := range e.Items {
-			names = append(names, extractRefNames(item)...)
-		}
-	case *Map:
-		for _, v := range e.KVs {
-			names = append(names, extractRefNames(v)...)
-		}
-	case *App:
-		for _, param := range e.Params {
-			names = append(names, extractRefNames(param)...)
-		}
-	}
-
-	return names
-}
 
 // Expr represents an expression in the AST.
 type Expr interface{}
